% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{check_background_job}
\alias{check_background_job}
\title{Check the status of a background job}
\usage{
check_background_job(job_id)
}
\arguments{
\item{job_id}{The job identifier to check}
}
\description{
Check the status of a background job
}
