% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{log_error_to_file}
\alias{log_error_to_file}
\title{Log error to file}
\usage{
log_error_to_file(code, error_message, log_path, agent_id = NULL)
}
\arguments{
\item{code}{The R code that caused the error}

\item{error_message}{The error message}

\item{log_path}{The path to the log file}
}
\value{
Invisible NULL
}
\description{
Log error to file
}
