% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_log.R
\name{stop_console_logging}
\alias{stop_console_logging}
\title{Stop logging the user's console activity}
\usage{
stop_console_logging()
}
\value{
Invisibly TRUE.
}
\description{
Stop logging the user's console activity
}
