% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{write_log_header}
\alias{write_log_header}
\title{Write reproducibility header to a new log file}
\usage{
write_log_header(log_path)
}
\arguments{
\item{log_path}{The path to the log file}
}
\value{
Invisible NULL
}
\description{
Captures sessionInfo(), working directory, and timestamp at the top of the log.
Called once when a new log file is created.
}
