% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{claude_rstudio_addin}
\alias{claude_rstudio_addin}
\title{Claude RStudio Connection Addin}
\usage{
claude_rstudio_addin(port = 8787)
}
\arguments{
\item{port}{the local port at which to run the server.}
}
\description{
Claude RStudio Connection Addin
}
