% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R, R/ui.R
\name{log_code_to_file}
\alias{log_code_to_file}
\title{Log code to file}
\usage{
log_code_to_file(code, log_path)

log_code_to_file(code, log_path)
}
\arguments{
\item{code}{The R code to log}

\item{log_path}{The path to the log file}
}
\value{
Invisible NULL

Invisible NULL
}
\description{
Log code to file

Log code to file
}
