% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{log_code_to_file}
\alias{log_code_to_file}
\title{Log code to file}
\usage{
log_code_to_file(code, log_path, agent_id = NULL)
}
\arguments{
\item{code}{The R code to log}

\item{log_path}{The path to the log file}

\item{agent_id}{Optional agent identifier used to attribute the entry}
}
\value{
Invisible NULL
}
\description{
Log code to file
}
